Stamp (Image) Tool
The first edition of the image tool was created.
Right now the program takes a lot of processing power to render each
frame because it has to render each object each frame. In order to
combat that, I plan on implementing a system that will render only the
objects that change each frame, saving on flops.